requirements.txt
Django>=5.2.0
sentence-transformers>=2.0.0
transformers>=4.0.0
torch>=1.9.0
Pillow>=8.0.0
numpy>=1.21.0

